[fix]レイアウトを調整。アプリ本体をindexの方に移動
