+++ package.json
@@ -8,7 +8,9 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "format": "prettier --write --ignore-path .gitignore './**/*.{js,jsx,ts,tsx,json,css}'"
+    "export": "next export",
+    "format": "prettier --write --ignore-path .gitignore './**/*.{js,jsx,ts,tsx,json,css}'",
+    "serve": "serve ./out"
   },
   "dependencies": {
     "axios": "^0.21.4",
@@ -27,6 +29,7 @@
     "eslint-config-prettier": "^8.3.0",
     "jest": "^27.2.3",
     "prettier": "^2.4.1",
+    "serve": "^12.0.1",
     "ts-jest": "^27.0.5",
     "typescript": "4.4.3"
   },
